requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
alpaca-trade-api==3.2.0
attrs==25.3.0
certifi==2025.7.14
charset-normalizer==3.4.2
deprecation==2.1.0
frozenlist==1.7.0
idna==3.10
msgpack==1.0.3
multidict==6.6.3
numpy==2.3.1
packaging==25.0
pandas==2.3.1
propcache==0.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.1
requests==2.32.4
six==1.17.0
tzdata==2025.2
urllib3==1.26.20
websocket-client==1.8.0
websockets==10.4
yarl==1.20.1
